Using argc and argv in c

+++ 0x0A-argc_argv/3-mul.c
@@ -8,8 +8,8 @@
  */
 int main(int argc, char *argv[])
 {
-	int r;
-	int k, l;
+	long int r;
+	long int k, l;
 
 	if (argc > 1)
 	{
